Criação do CRUD do cadastro do questionário para testes com o JPA. Organização de alguns métodos e classes de testes.

+++ quest/src/main/java/br/com/cetaceo/controller/QuestController.java
@@ -1,17 +1,25 @@
 package br.com.cetaceo.controller;
 
+import java.net.URI;
 import java.util.List;
 
+import javax.transaction.Transactional;
+import javax.validation.Valid;
+
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.util.UriComponentsBuilder;
 
-import br.com.cetaceo.dto.BA2EduTesteDTO;
+import br.com.cetaceo.dto.BA2EduTesteDto;
+import br.com.cetaceo.dto.QuestDto;
 import br.com.cetaceo.form.QuestForm;
 import br.com.cetaceo.model.Quest;
 import br.com.cetaceo.service.QuestService;
@@ -22,50 +30,47 @@ public class QuestController {
 	@Autowired
 	private QuestService questService;
 	
-	@GetMapping("/")
-	public String oi() {
-		return "Oi, eu sou a API Quest";
-	}
-	
-	@GetMapping("/myendpoint")
-	public String myendpointPub() {
-		return "REST Quest público";
-	}
-	
-	@GetMapping("/a/myendpoint")
-	public String myendpointAut() {
-		return "REST Quest autenticado";
-	}
-	
 	@GetMapping("/a/questionarios")
 	public Iterable<Quest> getAll() {
 		return questService.getQuestionarios();
 	}
 	
-	@PostMapping("/a/questionario/")
-	public void getById(@RequestBody QuestForm questForm) {
-		questService.saveQuestionario(questForm);
-	}
-	
 	@GetMapping("/a/questionario/{id}")
-	public Quest getById(@PathVariable Long id) {
-		return questService.getQuestionario(id);
+	public ResponseEntity<Quest> getById(@PathVariable Long id) {
+		Quest quest = questService.getQuestionario(id);
+		if ( quest != null )
+			return ResponseEntity.ok(quest);
+		else
+			return ResponseEntity.notFound().build();
 	}
 	
+	@Transactional
+	@PostMapping("/a/questionario/")
+	public ResponseEntity<QuestDto> save(@RequestBody @Valid QuestForm questForm, UriComponentsBuilder uriBuilder) {
+		Quest quest = questService.saveQuestionario(questForm);
+		URI uri = uriBuilder.path("/a/questionario/{id}").buildAndExpand(quest.getId()).toUri();
+		return ResponseEntity.created(uri).body(new QuestDto(quest));
+	}
+	
+	@Transactional
+	@PutMapping("/a/questionario/{id}")
+	public ResponseEntity<QuestDto> update(@PathVariable Long id, @RequestBody @Valid QuestForm questForm) {
+		Quest quest = questService.getQuestionario(id);
+		if ( quest != null ) {
+			quest = questForm.update(id, questService);
+			return ResponseEntity.ok(new QuestDto(quest));
+		} else
+			return ResponseEntity.notFound().build();
+	}
+	
+	@Transactional
 	@DeleteMapping("/a/questionario/{id}")
 	public ResponseEntity<?> deleteById(@PathVariable Long id) {
-		questService.deleteQuestionario(id);
-		return ResponseEntity.ok().build();
-	}
-	
-	
-	@GetMapping(path = {"/ba2edu","/ba2edu/{value}"})
-	public BA2EduTesteDTO ba2EduPub(@PathVariable(required = false) String value) {
-		return questService.ba2EduPub(value);
-	}
-	
-	@GetMapping(path = {"/a/ba2edu","/a/ba2edu/{value}"})
-	public BA2EduTesteDTO ba2EduAut(@PathVariable(required = false) String value) {
-		return questService.ba2EduAut(value);
+		Quest quest = questService.getQuestionario(id);
+		if ( quest != null ) {
+			questService.deleteQuestionario(id);
+			return ResponseEntity.ok().build();
+		} else
+			return ResponseEntity.notFound().build();
 	}
 }
